requirements.txt
python-dotenv==1.0.0
SQLAlchemy==2.0.23
instagrapi==2.0.1
Pillow==10.1.0